+++ lib/ts/recipe/session/utils.ts
@@ -162,8 +162,24 @@ export function validateAndNormaliseUserInput(
         signOutFeature.disableDefaultImplementation = config.signOutFeature.disableDefaultImplementation;
     }
 
-    let enableAntiCsrf =
-        config === undefined || config.enableAntiCsrf === undefined ? cookieSameSite === "none" : config.enableAntiCsrf;
+    if (config !== undefined && config.antiCsrf !== undefined) {
+        if (config.antiCsrf !== "NONE" && config.antiCsrf !== "VIA_CUSTOM_HEADER" && config.antiCsrf !== "VIA_TOKEN") {
+            throw new STError(
+                {
+                    type: STError.GENERAL_ERROR,
+                    payload: new Error("antiCsrf config must be one of 'NONE' or 'VIA_CUSTOM_HEADER' or 'VIA_TOKEN'"),
+                },
+                recipeInstance
+            );
+        }
+    }
+
+    let antiCsrf: "VIA_TOKEN" | "VIA_CUSTOM_HEADER" | "NONE" =
+        config === undefined || config.antiCsrf === undefined
+            ? cookieSameSite === "none"
+                ? "VIA_CUSTOM_HEADER"
+                : "NONE"
+            : config.antiCsrf;
 
     let errorHandlers: NormalisedErrorHandlers = {
         onTokenTheftDetected: (
@@ -201,18 +217,6 @@ export function validateAndNormaliseUserInput(
         }
     }
 
-    if (cookieSameSite === "none" && !enableAntiCsrf) {
-        throw new STError(
-            {
-                type: STError.GENERAL_ERROR,
-                payload: new Error(
-                    'Security error: enableAntiCsrf can\'t be set to false if cookieSameSite value is "none".'
-                ),
-            },
-            recipeInstance
-        );
-    }
-
     if (
         cookieSameSite === "none" &&
         !cookieSecure &&
@@ -241,7 +245,7 @@ export function validateAndNormaliseUserInput(
         sessionExpiredStatusCode,
         sessionRefreshFeature,
         errorHandlers,
-        enableAntiCsrf,
+        antiCsrf,
         signOutFeature,
     };
 }