+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [5.0.0] - 2021-05-02
+### Breaking change:
+- Uses custom header as a measure to prevent CSRF attacks - unless the user explicitly enables anti-csrf tokens.
+- new FDI 1.8
+- changes `enableAntiCsrf` config to `antiCsrf`.
+- removes deprecated `apiWebProxyPath` config. Use `appInfo -> apiGatewayPath` instead.
+
 ## [4.4.1] - 2021-04-29
 ### Fixed:
 - Runs JSON body parser only if `req.body` is `undefined` or if it is buffer